+++ packages/mds-config-service/index.ts
@@ -1,1 +1,1 @@
-export { client, ConfigManager } from './client'
+export { client, ConfigurationManager } from './client'
