+++ packages/mds-trip-processor/src/configuration.ts
@@ -1,0 +1,12 @@
+import { ConfigurationManager } from '@mds-core/mds-config-service'
+
+interface TripProcessorConfig {
+  compliance_sla: {
+    max_telemetry_time: number
+    max_telemetry_distance: number
+  }
+}
+
+const manager = ConfigurationManager<TripProcessorConfig>(['compliance_sla'])
+
+export const getConfig = async () => manager.configuration()
